pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nutrition-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "discord-py>=2.7.1",
    "google-generativeai>=0.8.6",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
nutrition-bot = "nutrition_bot.main:run_bot"

[tool.hatch.build.targets.wheel]
packages = ["src/nutrition_bot"]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
